chore(cleanup): remove unused code
