+++ pkg/agent/state_test.go
@@ -3,7 +3,7 @@ package agent
 import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"stvz.io/coral/pkg/util"
+	stvziov1 "stvz.io/coral/pkg/apis/stvz.io/v1"
 )
 
 var _ = Describe("Agent", func() {
@@ -11,10 +11,10 @@ var _ = Describe("Agent", func() {
 		Context("UpdateState", func() {
 			It("should update the state when all are available", func() {
 				nodeImages := map[string]string{
-					"image1": util.HashedImageLabelKey("image1"),
+					"image1": stvziov1.HashedImageLabelKey("image1"),
 				}
 				managedImages := map[string]string{
-					"image1": util.HashedImageLabelKey("image1"),
+					"image1": stvziov1.HashedImageLabelKey("image1"),
 				}
 
 				labels := UpdateState(nodeImages, managedImages)
@@ -25,7 +25,7 @@ var _ = Describe("Agent", func() {
 			It("should set a label to pending if the image is not available", func() {
 				nodeImages := map[string]string{}
 				managedImages := map[string]string{
-					"image1": util.HashedImageLabelKey("image1"),
+					"image1": stvziov1.HashedImageLabelKey("image1"),
 				}
 
 				labels := UpdateState(nodeImages, managedImages)
@@ -37,9 +37,9 @@ var _ = Describe("Agent", func() {
 		Context("ReplaceImageLabels", func() {
 			It("should replace all of the image labels with the new labels", func() {
 				nodeLabels := map[string]string{
-					util.HashedImageLabelKey("image1"): "available",
-					"kubernetes.io/arch":               "arm64",
-					"kubernetes.io/os":                 "linux",
+					stvziov1.HashedImageLabelKey("image1"): "available",
+					"kubernetes.io/arch":                   "arm64",
+					"kubernetes.io/os":                     "linux",
 				}
 				state := map[string]string{
 					"image1": "available",
@@ -48,17 +48,17 @@ var _ = Describe("Agent", func() {
 
 				labels := ReplaceImageLabels(nodeLabels, state)
 				Expect(labels).To(HaveLen(4))
-				Expect(labels).To(HaveKeyWithValue(util.HashedImageLabelKey("image1"), "available"))
-				Expect(labels).To(HaveKeyWithValue(util.HashedImageLabelKey("image2"), "pending"))
+				Expect(labels).To(HaveKeyWithValue(stvziov1.HashedImageLabelKey("image1"), "available"))
+				Expect(labels).To(HaveKeyWithValue(stvziov1.HashedImageLabelKey("image2"), "pending"))
 				Expect(labels).To(HaveKeyWithValue("kubernetes.io/arch", "arm64"))
 				Expect(labels).To(HaveKeyWithValue("kubernetes.io/os", "linux"))
 			})
 
 			It("should remove any labels that are not managed", func() {
 				nodeLabels := map[string]string{
-					util.HashedImageLabelKey("image1"): "available",
-					"kubernetes.io/arch":               "arm64",
-					"kubernetes.io/os":                 "linux",
+					stvziov1.HashedImageLabelKey("image1"): "available",
+					"kubernetes.io/arch":                   "arm64",
+					"kubernetes.io/os":                     "linux",
 				}
 				state := map[string]string{}
