+++ pkg/util/helpers.go
@@ -1,15 +1,8 @@
 package util
 
-import (
-	"crypto/md5" // #nosec
-	"fmt"
-	"math/rand"
-)
+// #nosec
 
 // TODO: Make this configurable later.
-const (
-	LabelPrefix = "image.stvz.io/"
-)
 
 func ModifyFunc[S, T any](a []S, b []T, fn func(S, T) S) []S {
 	var result []S
@@ -45,31 +38,4 @@ func FilterMapFunc[S, T comparable](a map[S]T, fn func(S, T) bool) map[S]T {
 	}
 
 	return result
-}
-
-// TODO: integrate the hashing into this.
-func ImageLabelKey(hash string) string {
-	return fmt.Sprintf("%s%s", LabelPrefix, hash)
-}
-
-func HashedImageLabelKey(name string) string {
-	hash := ImageHasher(name)
-	return fmt.Sprintf("%s%s", LabelPrefix, hash)
-}
-
-func ImageHasher(name string) string {
-	hasher := md5.New() // #nosec
-	hasher.Write([]byte(name))
-	return fmt.Sprintf("%x", hasher.Sum(nil))
-}
-
-func RandString(n int) string {
-	b := make([]rune, n)
-	chars := []rune("abcdefghijklmnopqrstuvwxyz1234567890")
-
-	for i := range b {
-		b[i] = chars[rand.Intn(len(chars))] // #nosec
-	}
-
-	return string(b)
 }
