--- pkg/util/namespace.go
@@ -1,14 +1,0 @@
-package util
-
-import (
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-func GenerateTestNamespace() corev1.Namespace {
-	return corev1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "coral-" + RandString(8),
-		},
-	}
-}
